// src/app/core/models/auth.ts
import { Role } from "./role";

export class Auth {
  // id: number;
  // username: string;
  // email: string;
  // roles: [string];
  // accessToken: string;
  // tokenType: string;
 
 message: string;
    statusCode: number;
    employee: any;
   token: string;
   username: string;
   id:number;
    refreshToken: string;
  entity: any;
}